ver 1.0 or close enough
cleaned up some, implemented automation entity purchasing, and updating the visible values. Really nastily implemented some basic upgrades that rely way too much on magic numbers. Need to basically refactor all of this, but I have an hour before I have to turn this in, so I'm gonna keep testing and turn it in probably and then forget about it for a minute.

At least at the moment I think it's pretty much completely functional. I'm stopping for the time being given I have few minutes left and that's not enough time to complete a significant enough refactor or enough time to really comprehend and rebalance the place that it's at.